css/css-tables/width-distribution/assignable_table_width_redistribution.tentative.html

<!doctype html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<script src="/resources/check-layout-th.js"></script>
<link rel="author" title="Aleks Totic" href="atotic@chromium.org" />
<link rel="help" href="https://www.w3.org/TR/css-tables-3/#distributing-width-to-columns" />
<style>
  table {
    background: gray;
    border-spacing: 8px 8px;
  }
  td {
    padding: 0;
    background: #BFB;
    font-size: 10px;
  }
  td > div {
    display: inline-block;
    background: rgba(56,162,56,0.3);
  }
  .error {
    color: red;
  }
  p {
    margin-top:4px;
    margin-bottom:4px;
  }
  body {
    counter-reset: testid;      /* Create a chapter counter scope */
  }
  .testdesc {
    white-space: pre-line;
    margin-top: 16px;
  }
  .testdesc:before {
    content: counter(testid) ". ";
    counter-increment: testid;  /* Add 1 to chapter */
  }
  .testdesc:first-line {
    font-family: monospace;
  }
</style>
<h1>Compute column computed widths from assignable table width</h1>
<h2>Test design</h2>
<p>All examples have border-spacing:8, td.padding:0</p>

<h2>Auto columns distribution</h2>

<p class="testdesc">Assi:300px C0: Auto/75/75 C1:Auto/25/25
Non-empty auto cells get surplus width proportionally to their max width.
Guess3: 100. Guess4: 300, diff 200, fixed priority.
C0: 75 + 75/100*200 = 225  C1: 25 + 25/100*200 = 75</p>
<table style="width:calc(300px + 24px)" data-expected-width=324>
  <tr>
    <td data-expected-width=225><div style="width:75px">75</div></td>
    <td data-expected-width=75><div style="width:25px">25</div></td>
  </tr>
</table>
<p class="testdesc">Assignable:300px C0: Auto/75/75 C1:Auto/13/25 C2:Auto/0/0
Empty cells get nothing if there are non-empty auto cells.
Guess3: 100, Guess4: 300; diff 200, fixed priority.
C0: 75 + 75/100*200 = 225  C1: 25 + 25/100*200 = 75; C2: 0
</p>
<table style="width:calc(300px + 32px)" data-expected-width=332>
  <tr>
    <td data-expected-width=225>
      <div style="width:75px">75</div></td>
    <td data-expected-width=75>
      <div style="width:13px">13</div><div style="width:12px">12</div></td>
    <td data-expected-width=0></td>
  </tr>
</table>

<h2>Guess 0: Auto(min), Fixed(min), Percentage(min) to Guess 1.</h2>

<p class="testdesc">Assi: 1px; C0:Auto/50/100 C1:100/50/100 C2:50%/50/100
All columns get minimum width.
Guess0: 150
C0: 50, C1:50, C2: 50</p>
<table style="width:1px" data-expected-width=182>
  <tr>
    <td data-expected-width=50>
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=50 style="width:100px">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=50 style="width:50%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
</table>

<p class="testdesc">Assi: 160px; C0:Auto/50/100 C1:100/50/100 C2:50%/50/100
%ge column grows.
Guess0: 150, Guess1: 180; diff 10.
C0: 50, C1:50, C2: 50 + 10 * 10/10 = 60</p>
<table style="width:calc(160px + 32px)" data-expected-width=192>
  <tr>
    <td data-expected-width=50>
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=50 style="width:100px">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=60 style="width:50%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
</table>

<p class="testdesc">Assi: 210px; C0:Auto/50/100 C1:100/50/100 C2:30%/50/100 C3:30%/50/100
2 %ge columns grow evenly.
Guess 0: 200, Guess 1: 240, diff 10
C2: 50 + 10 * 70/140 C3: 50 + 10 * 70/140
<table style="width:calc(40px + 210px)" data-expected-width=250>
  <tr>
    <td data-expected-width=50>
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=50 style="width:100px">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=55 style="width:30%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=55 style="width:30%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
</table>

<p class="testdesc">Assi: 220px; C0:Auto/50/100 C1:100/50/100 C2:25%/50/100 C3:40%50/100
%ge columns grow in proportion to increase from previous guess.
Guess 0: 200. C2[G1] = .25*220 = 55 C3[G1] = .4 * 220 = 88 Guess 1: 100 + 55 + 88 = 243
diff = 220 - 200 = 20. C2 grew 5, C3 grew 38, total grew 43.
C2: 50 + 20 *5/43 = 52.32; C3: 50 + 20 * 38 / 43 = 67.67
<table style="width:calc(40px + 220px)" data-expected-width=260>
  <tr>
    <td data-expected-width=50>
      <div style="width:50px">50</div> <div style="width:50px">50</div></td>
    <td data-expected-width=50 style="width:100px">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width=52 style="width:25%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td  data-expected-width=68 style="width:40%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
</table>

<h2>Guess 1 to Guess 2: Auto(min), Percentage(%max) to Fixed(min => max)</h2>
<p>These tests are non-intuitive to evaluate. When table size increases betwee Guess 1 and 2,
  although the standard says that fixed columns are growing, %ge columns grow too because their max width depends on table width.</p>

<p class="testdesc">Assi:166, C0:Auto/50/100 C1:100/50/100 C2:40%/50
Edge example, Guess 1 %ge cell has grown to the max.
C2: grows to .4*165 = 66.4, table is 166.4+32 = 198.4</p>
<table style="width:calc(166px + 32px)" data-expected-width=198>
  <tr>
    <td data-expected-width="50" >
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="50" style="width:100px">
       <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="66" style="width:40%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
  </tr>
</table>

<p class="testdesc">Assi:216, C0:Auto/50/100 C1:100/50/100 C2:40%/50
  %ge cell grows to the max, the rest goes to fixed.
Guess 1 size is 50 + 50 + (.4*216=>86.4) = 186.4
Guess 2 size is 50 + 100 + 86.4 = 236
Assi - Guess1 = 29.6; C2 = 50 + 29.6 * 50/50 = 79.6
</p>
<table style="width:calc(216px + 32px)" data-expected-width=248>
  <tr>
    <td data-expected-width="50" >
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="80" style="width:100px">
       <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="86" style="width:40%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
  </tr>
</table>

<h2>Guess 2 to Guess 3: Percentage(%max), Fixed(max), Auto(min=>max)</h2>

<p class="testdesc">Assi:300
Guess 2 size is 50 + 100 + .4*300 = 270
Guess 3 size is 100 + 100 + 120 = 320
Assi - Guess2 = 30, C0 = 50 + 3- = 80
<table style="width:calc(300px + 32px)" data-expected-width=332>
  <tr>
    <td data-expected-width="80" >
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="100" style="width:100px">
       <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="120" style="width:40%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
  </tr>
</table>

<h2>Guess3 to Guess4, Auto(max), Percentage(%max), Fixed(max) grow first available Auto|Fixed|Percent</h2>

<p class="testdesc">Assi: 500, C0:Auto, C1: Fixed, C2: Percent
Guess 3 is 100 + 100 + .4 * 500 = 400
C0 gets the 100
<table style="width:calc(500px + 32px)" data-expected-width=532>
  <tr>
    <td data-expected-width="200" >
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="100" style="width:100px">
       <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="200" style="width:40%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
  </tr>
</table>

<p class="testdesc">Assi: 500, C0:Fixed, C1: Fixed, C2: Percent
Guess 3 is 100 + 100 + .4 * 500 = 400, 100 to be redistributed
Fixed cells, C0 and C1 get 50 each.
<table style="width:calc(500px + 32px)" data-expected-width=532>
  <tr>
    <td data-expected-width="150" style="width:100px">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="150" style="width:100px">
       <div style="width:50px">50</div><div style="width:50px">50</div></td>
    <td data-expected-width="200" style="width:40%">
      <div style="width:50px">50</div><div style="width:50px">50</div></td>
  </tr>
</table>

<p class="testdesc">Assi: 700, C0:10%/40, C1: 20%/50, C2: 40%/50, C3: 100%/50
Percentage cells only.
Compute columns as %ge of total width:
C0: 700*10/70, C1: 700*20/70 C2: 700*30/70
<table style="width:calc(700px + 32px)" data-expected-width=732>
  <tr>
    <td data-expected-width=100 style="width:10%">
      <div style="width:40px">40</div></td>
    <td data-expected-width=200 style="width:20%">
       <div style="width:50px">50</div></td>
    <td data-expected-width=400 style="width:40%">
      <div style="width:50px">50</div></td>
  </tr>
</table>
<p class="testdesc">Assi: 600, C0:10%/40, C1: 20%/50, C2: 40%/50, C3: 100%/50
Percentage cells only. Over 100% columns get their percentage truncated.
<table style="width:calc(600px + 40px)" data-expected-width=640>
  <tr>
    <td data-expected-width=60 style="width:10%">
      <div style="width:40px">40</div></td>
    <td data-expected-width=120 style="width:20%">
       <div style="width:50px">50</div></td>
    <td data-expected-width=240 style="width:40%">
      <div style="width:50px">50</div></td>
    <td data-expected-width=180 style="width:100%">
      <div style="width:50px">50</div></td>
  </tr>
</table>

<script>
// Table inspector. Annotates cells with table measures.
function measureCellMinMax(cell) {
  let d = document.createElement("div");
  let clone = cell.cloneNode(true);
  d.classList.add("measure");
  for (child of Array.from(clone.childNodes))
    d.appendChild(child);
  d.style.width = "min-content";
  document.body.appendChild(d);
  let min = d.offsetWidth;
  d.style.width = "max-content";
  let max = d.offsetWidth;
  d.remove();
  return {min: min, max: max};
}
function annotateTable(t) {
  let tableWidth = t.offsetWidth;
  let inlineBorderSpacing = parseInt(window.getComputedStyle(t).borderSpacing.split(' ')[0]);
  let spacing = (2 + t.querySelector("tr").querySelectorAll("td").length - 1)*inlineBorderSpacing;
  t.setAttribute("title", `Table: ${tableWidth.toFixed(0)} spacing: ${spacing}px`);
  let firstCell;
  let totalCellPercent = 0;
  let nonPercentCellMinWidth = 0;
  let nonPercentCellMaxWidth = 0;
  let tableMinWidthByCellPercent = 0;
  // Spacing assumes column count specified by the 1st row. Assumption might be wrong.
  let cells = Array.from(t.querySelectorAll("td"));
  for (let cell of cells) {
    if (!firstCell)
      firstCell = cell;
    let percent = cell.offsetWidth / (tableWidth - spacing) * 100;
    let minmax = measureCellMinMax(cell);
    let title = `${cell.offsetWidth.toFixed(1)}px\nmin:${minmax.min}px max:${minmax.max}px ${percent.toFixed(1)}%`;
    let cssWidth = cell.style.width;
    let is_percent = cssWidth && cssWidth.match(/\%/);
    if (is_percent) {
      let w = parseFloat(cssWidth);
      totalCellPercent += w;
      let tableMinWidth = minmax.min / (w / 100);
      tableMinWidthByCellPercent = Math.max(tableMinWidthByCellPercent, tableMinWidth);
      title += `\nmin table: ${tableMinWidth.toFixed(0)}px`;
    } else {
      nonPercentCellMinWidth += minmax.min;
      nonPercentCellMaxWidth += minmax.max;
    }
    title += `\nTable: ${tableWidth.toFixed(0)} spacing: ${spacing}px`;
    cell.setAttribute("title", title);
  }


  // Display table statistics in first cell.
  if (firstCell) {
    let title = firstCell.getAttribute("title");
    let ruleMatch = 0;
    if (tableMinWidthByCellPercent != 0) {
      if ((tableMinWidthByCellPercent + spacing) == tableWidth)
        ruleMatch += 1;
      title += `\nTable by rule 1 min : ${tableMinWidthByCellPercent.toFixed(1)}`;
    } else {
      title += "\nTable min by single cell percent NA";
    }
    if (totalCellPercent > 0) {
      totalCellPercent = Math.min(totalCellPercent, 100);
      let minByCellPercent
    }
    if (nonPercentCellMinWidth && totalCellPercent > 0) {
      totalCellPercent = Math.min(totalCellPercent, 100);
      title += `\nsum%: ${totalCellPercent.toFixed(1)}%; non% min: ${nonPercentCellMinWidth}px; non% max ${nonPercentCellMaxWidth}px`;
      let tableMinBySum = (totalCellPercent / (100 - totalCellPercent) +1) * nonPercentCellMinWidth;
      let tableMaxBySum = (totalCellPercent / (100 - totalCellPercent) +1) * nonPercentCellMaxWidth;
      if (Math.floor((tableMinBySum + spacing)) == Math.floor(tableWidth) ||
        Math.floor((tableMaxBySum + spacing)) == Math.floor(tableWidth))
        ruleMatch += 2;
      title += `\nTable by rule 2 ${totalCellPercent}%; min:${tableMinBySum.toFixed(1)}px; max:${tableMaxBySum.toFixed(1)}px;`;
    } else {
      "Table min by % sum not available";
    }
    firstCell.setAttribute("title", title);
    switch(ruleMatch) {
      case 1: t.classList.toggle('rule1'); break;
      case 2: t.classList.toggle('rule2'); break;
      case 3: t.classList.toggle('rule1and2'); break;
      default: break;
    }
  }
}
for (let t of Array.from(document.querySelectorAll("table")))
  annotateTable(t);
checkLayout("table");
</script>
